Fix the Catalyst build after recent changes in UIFoundation (rdar://164301290) https://bugs.webkit.org/show_bug.cgi?id=303488 rdar://165774936
Reviewed by Aditya Keerthi and Abrar Rahman Protyasha.

Fix the build after the changes in rdar://164301290; we avoid importing `<UIFoundation/NSTextTable.h>`
on Catalyst, since the header does not declare `NSTextBlock` when `TARGET_OS_MACCATALYST` is defined
(so we need to fall back on forward declarations).

* Source/WebCore/PAL/pal/spi/ios/UIKitSPI.h:
* Source/WebCore/editing/cocoa/AttributedString.mm:
(WebCore::extractParagraphStyle):
* Source/WebCore/editing/cocoa/NodeHTMLConverter.mm:
(HTMLConverter::_exitElement):

Canonical link: https://commits.webkit.org/303862@main

+++ Source/WebCore/editing/cocoa/AttributedString.mm
@@ -619,7 +619,7 @@ inline static ParagraphStyle extractParagraphStyle(NSParagraphStyle *style, Tabl
                     [nsTable borderColorForEdge:NSRectEdgeMaxY]
                 },
                 tableID,
-                [nsTable numberOfColumns],
+                static_cast<uint64_t>([nsTable numberOfColumns]),
                 extractTextTableLayoutAlgorithm([nsTable layoutAlgorithm]),
                 !![nsTable collapsesBorders],
                 !![nsTable hidesEmptyCells],
